+++ CHANGELOG.md
@@ -5,10 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.1.1
+## Unreleased
+### Added
+- Some histograms for 3B femto
+- Different functions for computing the mT of a triplet
+- Add option to set the arbitrary mass for 3B calculations
 
+## 0.1.1
 ### Added
 - CHANGELOG.md file
-
+- some histograms in CECA for 3B femto
 ### Fixed
 - Use Q3 instead of Q in 3B calculations
+
+## 0.1.0
+### Changed
+- Use cached user paths during installation
+### Added
+- Some basic tests
+### Fixed
+- Compilation errors